% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{partyscape_palette}
\alias{partyscape_palette}
\title{partyscape palette}
\usage{
partyscape_palette(n = 8L)
}
\arguments{
\item{n}{number of colours to return (recycled if n > length of palette).}
}
\value{
character vector of hex codes.
}
\description{
A small qualitative palette used by the autoplot methods.
}
