% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.beta_diversity}
\alias{plot.beta_diversity}
\title{Plot a beta_diversity object}
\usage{
\method{plot}{beta_diversity}(x, ...)
}
\arguments{
\item{x}{a \code{beta_diversity} object.}

\item{...}{forwarded to \code{\link[=autoplot.diversity_profile]{autoplot.diversity_profile()}}.}
}
\description{
Plot a beta_diversity object
}
