% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.crossings}
\alias{plot.crossings}
\title{Plot a crossings object}
\usage{
\method{plot}{crossings}(x, ...)
}
\arguments{
\item{x}{a \code{crossings} object.}

\item{...}{forwarded to \code{\link[=autoplot.diversity_profile]{autoplot.diversity_profile()}}.}
}
\description{
Plot a crossings object
}
