% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.evenness_profile}
\alias{plot.evenness_profile}
\title{Plot an evenness_profile}
\usage{
\method{plot}{evenness_profile}(x, ...)
}
\arguments{
\item{x}{an \code{evenness_profile} object.}

\item{...}{forwarded to \code{\link[=autoplot.diversity_profile]{autoplot.diversity_profile()}}.}
}
\description{
Plot an evenness_profile
}
