% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalars.R
\name{rae}
\alias{rae}
\title{Rae's fractionalization}
\usage{
rae(p)
}
\arguments{
\item{p}{numeric vector / matrix / data.frame of shares.}
}
\description{
\code{rae(p) = 1 - hhi(p)}. The probability that two randomly chosen parties
(with replacement, by share) come from different parties.
}
