% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{[.diversity_profile}
\alias{[.diversity_profile}
\title{Subset a diversity profile by id}
\usage{
\method{[}{diversity_profile}(x, i, ...)
}
\arguments{
\item{x}{a \code{diversity_profile} object.}

\item{i}{character or integer vector of ids to keep.}

\item{...}{ignored.}
}
\description{
Subset a diversity profile by id
}
