% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalars.R
\name{top_two}
\alias{top_two}
\title{Top-two share}
\usage{
top_two(p)
}
\arguments{
\item{p}{numeric vector / matrix / data.frame of shares.}
}
\description{
Sum of the two largest shares. Reports the "two-big-parties" footprint.
}
